+++ x-pack/plugins/siem/public/store/model.ts
@@ -9,15 +9,4 @@ export { dragAndDropModel } from './drag_and_drop';
 export { hostsModel } from './hosts';
 export { inputsModel } from './inputs';
 export { networkModel } from './network';
-
-export type KueryFilterQueryKind = 'kuery' | 'lucene';
-
-export interface KueryFilterQuery {
-  kind: KueryFilterQueryKind;
-  expression: string;
-}
-
-export interface SerializedFilterQuery {
-  kuery: KueryFilterQuery | null;
-  serializedQuery: string;
-}
+export * from './types';
